+++ ref_data/retrieve_stips_data.py
@@ -3,7 +3,7 @@
 #    the same file (making sure to update 1.x.x version of Pandeia you're using
 #    –– for STIPS this should be 1.7):
 #       export stips_data="<absolute_path_to_this_folder>/ref_data/stips_data"
-#       export WEBBPSF_PATH="<absolute_path_to_this_folder>/ref_data/webbpsf-data"
+#       export STPSF_PATH="<absolute_path_to_this_folder>/ref_data/stpsf-data"
 #       export PYSYN_CDBS="<absolute_path_to_this_folder>/ref_data/grp/redcat/trds"
 #       export pandeia_refdata="<absolute_path_to_this_folder>/ref_data/pandeia_data-
 #       2024.12-roman"